Refactoring postFunctions
