+++ src/components/Foro/Modal.jsx
@@ -3,14 +3,14 @@ import { useState } from "react";
 import { useEffect } from "react";
 import { AiOutlinePlayCircle } from "react-icons/ai";
 import { RiImageFill } from "react-icons/ri";
-import { createNewPost } from "@/api/apiCall/functions";
 import { CustomContext } from "@/store/ContextProvider";
 import Image from "next/image";
 import Alerts from "../Alerts/Alerts";
+import { createNewPost } from "@/api/apiCall/PostFunctions";
 
 export default function ModalnewPost({ setvisible }) {
   const { UserContext } = CustomContext();
-  const [seeAlert, setSeeAlert] = useState(true)
+  const [seeAlert, setSeeAlert] = useState(true);
   useEffect(() => {
     const body = document.getElementById("Body");
     body && (body.style.overflow = "hidden");
@@ -74,7 +74,6 @@ export default function ModalnewPost({ setvisible }) {
       } catch (error) {
         console.error("Error:", error);
       }
-
     }
   };
 
@@ -93,11 +92,18 @@ export default function ModalnewPost({ setvisible }) {
     imagePreview = URL.createObjectURL(newPost.image);
   }
   const closeAlert = () => {
-    setSeeAlert(false)
-  }
+    setSeeAlert(false);
+  };
   return (
     <main className="bg-[#686868cc] z-10 min-h-screen fixed h-full w-full flex justify-center items-center top-0 overflow-scroll">
-      {seeAlert && <Alerts title={'Exito!'} textdetails={'El post fue subido correctamente'} closemodal={closeAlert} callback={setvisible} />}
+      {seeAlert && (
+        <Alerts
+          title={"Exito!"}
+          textdetails={"El post fue subido correctamente"}
+          closemodal={closeAlert}
+          callback={setvisible}
+        />
+      )}
       <div className="relative font-semibold h-[35rem] max-h-[90%] w-[40rem] max-w-[97%] bg-[#D9D9D9] flex flex-col justify-between md:h-[28rem] md:w-[38rem] lg:h-[30rem] lg:w-[40rem] xl:h-[33rem] xl:w-[43rem] 2xl:h-[35rem] 2xl:w-[45rem]">
         <button
           className="absolute right-0 top-0 text-2x1"
